+++ src/Service/AbstractJiraService.php
@@ -539,8 +539,8 @@ abstract class AbstractJiraService
                 'scopeType' => $data['scope_type'],
                 'scopeId' => $data['scope_id'],
             ],
-            'amount' => (int) ($data['quantity'] * $category->getUnitPrice()),
-            'description' => $data['description'],
+            'amount' => round($data['quantity'] * $category->getUnitPrice(), 2),
+            'description' => $data['description'] ?? $category->getName(),
             'date' => (new \DateTime())->format(\DateTime::ATOM),
         ];
         $result = $this->post('/rest/tempo-core/1/expense/', $data);