+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## in develop
 
 * Changed spreadsheet export to use temporary files.
+* Expense: Made description optional.
+* Expense: Added select2 to selects.
+* Expense: Use previous project selection
 
 ## 1.5.0